anv: Never copy dwords for clear color on gen10.
Since we already emit the surface state with a pointer to where the
color is stored, we can avoid copying the clear color from the surface
state buffer to the clear state entry buffer and vice-versa. Just store
the color on teh clear state entry when setting up the attachments.

Signed-off-by: Rafael Antognolli <rafael.antognolli@intel.com>